+++ libs/utils/source/ftrack_utils/calls/__init__.py
@@ -1,0 +1,4 @@
+# :coding: utf-8
+# :copyright: Copyright (c) 2024 ftrack
+
+from ftrack_utils.calls.methods import call_directly