+++ libs/framework-core/source/ftrack_framework_core/event/__init__.py
@@ -9,37 +9,9 @@ import ftrack_constants.framework as constants
 import uuid
 import time
 
+from ftrack_utils.event_hub import EventHubThread
+
 logger = logging.getLogger(__name__)
-
-
-class _EventHubThread(threading.Thread):
-    '''Listen for events from ftrack's event hub.'''
-
-    def __repr__(self):
-        return "<{0}:{1}>".format(self.__class__.__name__, self.name)
-
-    def __init__(self, session):
-        self.logger = logging.getLogger(
-            __name__ + '.' + self.__class__.__name__
-        )
-        _name = str(hash(session))
-        super(_EventHubThread, self).__init__(name=_name)
-        self.logger.debug('Name set for the thread: {}'.format(_name))
-        self._session = session
-
-    def start(self):
-        '''Start thread for *_session*.'''
-        self.logger.debug(
-            'starting event hub thread for session {}'.format(self._session)
-        )
-        super(_EventHubThread, self).start()
-
-    def run(self):
-        '''Listen for events.'''
-        self.logger.debug(
-            'hub thread started for session {}'.format(self._session)
-        )
-        self._session.event_hub.wait()
 
 
 class EventManager(object):
@@ -96,16 +68,18 @@ class EventManager(object):
             self.session.event_hub.connect()
 
     def _wait(self):
+        # Check if already has an event hub otherwise create one
         for thread in threading.enumerate():
-            if thread.getName() == str(hash(self.session)):
-                self._event_hub_thread = thread
-                break
+            if (
+                isinstance(thread, EventHubThread)
+                and thread._session == self.session
+            ):
+                if thread.name == str(hash(self.session)):
+                    self._event_hub_thread = thread
+                    break
         if not self._event_hub_thread:
-            # self.logger.debug('Initializing new hub thread {}'.format(self))
-            self._event_hub_thread = _EventHubThread(self.session)
-
+            self._event_hub_thread = EventHubThread(self.session)
         if not self._event_hub_thread.is_alive():
-            # self.logger.debug('Starting new hub thread for {}'.format(self))
             self._event_hub_thread.start()
 
     def __init__(self, session, mode=constants.event.LOCAL_EVENT_MODE):
@@ -502,29 +476,5 @@ class Subscribe(object):
         '''
         event_topic = '{} and data.host_id={}'.format(
             constants.event.HOST_VERIFY_PLUGINS_TOPIC, host_id
-        )
-        return self._subscribe_event(event_topic, callback)
-
-    def ftrack_action_discover(self, callback=None):
-        '''
-        Subscribe to an event with topic
-        :const:`~ftrack_framework_core.constants.event.FTRACK_ACTION_DISCOVER_TOPIC`
-        '''
-        event_topic = '{} and source.user.username={}'.format(
-            constants.event.FTRACK_ACTION_DISCOVER_TOPIC,
-            self.event_manager.session.api_user,
-        )
-        return self._subscribe_event(event_topic, callback)
-
-    def ftrack_action_launch(self, host_id, action_name, callback=None):
-        '''
-        Subscribe to an event with topic
-        :const:`~ftrack_framework_core.constants.event.FTRACK_ACTION_LAUNCH_TOPIC`
-        '''
-        event_topic = '{} and data.name={} and source.user.username={} and data.host_id={}'.format(
-            constants.event.FTRACK_ACTION_LAUNCH_TOPIC,
-            action_name,
-            self.event_manager.session.api_user,
-            host_id,
         )
         return self._subscribe_event(event_topic, callback)