+++ libs/utils/release_notes.md
@@ -1,11 +1,19 @@
 # ftrack Utils library release Notes
 
 
+## upcoming
+
+* [new] Calls methids; call_directly utility function added to directly call a function with the give arguments.
+* [new] Decorators threading; delegate_to_main_thread_wrapper added.
+
+
 ## v2.4.0rc1
 2024-07-17
 
+* [new] decorators; run_in_main_thread decorator added.
 * [fix] JS RPC; Properly pick up and handle error messages from DCC.
 * [changed] get_temp_path; Support temp directories.
+
 
 ## v2.3.0
 2024-06-04