+++ libs/utils/release_notes.md
@@ -4,6 +4,7 @@
 ## upcoming
 
 * [new] Actions, remote_actions; Add support to register remote actions.
+* [new] event_hub event_hub_thread; Added EventHubThread utility.
 * [new] Calls methods; call_directly utility function added to directly call a function with the give arguments.
 * [new] Decorators threading; delegate_to_main_thread_wrapper added.